requirements.txt
langchain>=0.3.0
langchain_core==0.3.76
langchain_openai>=0.1.0
langchain_community>=0.3.0
requests==2.32.5
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0

# Optional: for vLLM support (HuggingFace inference)
# vllm>=0.4.0
